BUG-26120 Phoenix client should clean up the temp files at /var/tmp after query completion

+++ phoenix-core/src/main/java/org/apache/phoenix/jdbc/PhoenixConnection.java
@@ -384,7 +384,8 @@ public class PhoenixConnection implements Connection, org.apache.phoenix.jdbc.Jd
             	List<File> tmpFiles = context.getTmpFiles();
             	if (tmpFiles != null) {
             		for (File file : tmpFiles) 
-            			file.deleteOnExit();
+            		    if (file != null)
+            		        file.deleteOnExit();
             	}
             }
             stmt.close();
